backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
bcrypt==4.1.2

# HTTP & Requests
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Data validation
pydantic==2.5.0
email-validator==2.1.0

# CORS
fastapi-cors==0.0.6

# Environment & Config
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# JSON Web Tokens
PyJWT==2.8.0

# Template & Export
Jinja2==3.1.2
python-slugify==8.0.1